Move orders to standalone component
