+++ src/components/ExchangePage/index.jsx
@@ -6,30 +6,23 @@ import * as ls from 'local-storage'
 import { ethers } from 'ethers'
 import styled from 'styled-components'
 
-import { isAddress } from '../../utils'
 import { Button } from '../../theme'
 import CurrencyInputPanel from '../CurrencyInputPanel'
-import { OrderCard } from '../OrderCard'
-import { OrdersHistory } from '../OrdersHistory'
 import OversizedPanel from '../OversizedPanel'
 import ArrowDown from '../../assets/svg/SVGArrowDown'
-import Circle from '../../assets/images/circle.svg'
 import SVGClose from '../../assets/svg/SVGClose'
 import SVGDiv from '../../assets/svg/SVGDiv'
 import { amountFormatter } from '../../utils'
-import { useUniswapExContract, useMulticallContract } from '../../hooks'
-import { Spinner } from '../../theme'
+import { useUniswapExContract } from '../../hooks'
 import { useTokenDetails } from '../../contexts/Tokens'
 import {
   ACTION_PLACE_ORDER,
-  useAllPendingOrders,
-  useTransactionAdder,
-  useAllPendingCancelOrders
+  useTransactionAdder
 } from '../../contexts/Transactions'
 import { useAddressBalance } from '../../contexts/Balances'
 import { useFetchAllBalances } from '../../contexts/AllBalances'
 import { useTradeExactIn } from '../../hooks/trade'
-import { ETH_ADDRESS, LIMIT_ORDER_MODULE_ADDRESSES, ORDER_GRAPH } from '../../constants'
+import { ETH_ADDRESS, LIMIT_ORDER_MODULE_ADDRESSES } from '../../constants'
 import { getExchangeRate } from '../../utils/rate'
 
 import './ExchangePage.css'
@@ -106,10 +99,6 @@ const Flex = styled.div`
   }
 `
 
-const SpinnerWrapper = styled(Spinner)`
-  margin: 0 0.25rem 0 0.25rem;
-`
-
 // ///
 // Local storage
 // ///
@@ -133,14 +122,6 @@ function saveOrder(account, orderData, chainId) {
       ls.set(key, prev)
     }
   }
-}
-
-function getSavedOrders(account, chainId) {
-  if (!account) return []
-
-  console.log('Loading saved orders from storage location', account, lsKey(LS_ORDERS, account, chainId))
-  const raw = ls.get(lsKey(LS_ORDERS, account, chainId))
-  return raw == null ? [] : raw
 }
 
 // ///
@@ -295,129 +276,6 @@ function flipRate(rate) {
   } catch {}
 }
 
-async function fetchUserOrders(account, chainId) {
-  const query = `
-  query GetOrdersByOwner($owner: String) {
-    orders(where:{owner:$owner,status:open}) {
-      id
-      owner
-      module
-      inputToken
-      outputToken
-      inputAmount
-      minReturn
-      witness
-      secret
-      status
-    }
-  }`
-  try {
-    const res = await fetch(ORDER_GRAPH[chainId], {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ query, variables: { owner: account.toLowerCase() } })
-    })
-
-    const { data } = await res.json()
-    return {
-      allOrders: [],
-      openOrders: data.orders
-    }
-  } catch (e) {
-    console.warn('Error loading orders from TheGraph', e)
-    return {
-      allOrders: [],
-      openOrders: []
-    }
-  }
-}
-
-function useGraphOrders(account, chainId) {
-  const [state, setState] = useState({ openOrders: [], allOrders: [] })
-
-  useEffect(() => {
-    console.log(`Requesting load orders from the graph`)
-    if (account && isAddress(account)) {
-      fetchUserOrders(account, chainId).then(orders => {
-        console.log(`Fetched ${orders.allOrders.length} ${orders.openOrders.length} orders from the graph`)
-        setState(orders)
-      })
-    }
-  }, [account, chainId])
-
-  return state
-}
-
-function isEthOrder(order) {
-  return order.inputToken.toLowerCase() === ETH_ADDRESS.toLowerCase()
-}
-
-function keyOfOrder(order) {
-  const moduleData = ethers.utils.defaultAbiCoder.encode(['address', 'uint256'], [order.outputToken, order.minReturn])
-
-  return ethers.utils.keccak256(
-    ethers.utils.defaultAbiCoder.encode(
-      ['address', 'address', 'address', 'address', 'bytes'],
-      [order.module, order.inputToken, order.owner, order.witness, moduleData]
-    )
-  )
-}
-
-function vaultForOrder(order, uniswapEXContract) {
-  const VAULT_CODE_HASH = '0xfa3da1081bc86587310fce8f3a5309785fc567b9b20875900cb289302d6bfa97'
-  const hash = ethers.utils.keccak256(
-    ethers.utils.solidityPack(
-      ['bytes1', 'address', 'bytes32', 'bytes32'],
-      ['0xff', uniswapEXContract.address, keyOfOrder(order), VAULT_CODE_HASH]
-    )
-  )
-
-  return `0x${hash.slice(-40)}`
-}
-
-async function balancesOfOrders(orders, uniswapEXContract, multicallContract) {
-  const result = await multicallContract.aggregate(
-    orders.map(o => {
-      if (!isEthOrder(o)) {
-        return [
-          o.inputToken,
-          `0x70a08231${ethers.utils.defaultAbiCoder // balanceOf(address)
-            .encode(['address'], [vaultForOrder(o, uniswapEXContract)])
-            .replace('0x', '')}`
-        ]
-      } else {
-        return [uniswapEXContract.address, `0xebd9c39c${keyOfOrder(o).replace('0x', '')}`] // ethDeposits(bytes32)
-      }
-    })
-  )
-
-  return result.returnData
-}
-
-function useSavedOrders(account, chainId, uniswapEXContract, multicallContract, deps = []) {
-  const [state, setState] = useState({ allOrders: [], openOrders: [] })
-
-  useEffect(() => {
-    console.log(`Requesting load orders from storage`)
-    if (isAddress(account)) {
-      const allOrders = getSavedOrders(account, chainId)
-      console.log(`Loaded ${allOrders.length} orders from local storage`)
-      if (allOrders.length > 0) {
-        balancesOfOrders(allOrders, uniswapEXContract, multicallContract).then(amounts => {
-          allOrders.map((o, i) => (o.inputAmount = ethers.utils.bigNumberify(amounts[i]).toString()))
-          setState({
-            allOrders: allOrders,
-            openOrders: allOrders.filter(o => o.inputAmount !== '0')
-          })
-        })
-      }
-    }
-    // eslint-disable-next-line
-  }, [...deps, account, chainId, uniswapEXContract])
-
-  return state
-}
-
 function safeParseUnits(number, units) {
   try {
     return ethers.utils.parseUnits(number, units)
@@ -438,34 +296,9 @@ export default function ExchangePage({ initialCurrency }) {
   const { independentValue, independentField, inputCurrency, outputCurrency, rateOp, inputRateValue } = swapState
 
   const uniswapEXContract = useUniswapExContract()
-  const multicallContract = useMulticallContract()
   const [inputError, setInputError] = useState()
 
-  const loading = !account
-
-  const pendingOrders = useAllPendingOrders()
-  const pendingCancelOrders = useAllPendingCancelOrders()
-
-  // Get locally saved orders and the graph orders
-  const local = useSavedOrders(account, chainId, uniswapEXContract, multicallContract, [
-    pendingOrders.length,
-    pendingCancelOrders.length
-  ])
-  const graph = useGraphOrders(account, chainId)
-
-  // Aggregate graph and local orders, local orders have priority
-  const allOrders = local.allOrders.concat(
-    graph.allOrders.filter(o => !local.allOrders.find(c => c.secret === o.secret))
-  )
-  const openOrders = local.openOrders.concat(
-    graph.openOrders.filter(o => !local.allOrders.find(c => c.secret === o.secret))
-  )
-
-  // Define orders to show as openOrders + pending orders
-  const orders = openOrders.concat(allOrders.filter(o => pendingOrders.find(p => p.secret === o.secret)))
-  // const orders = graph.openOrders.concat(knownOrders.filter(o => o && pendingOrders.indexOf(o) !== -1))
   const addTransaction = useTransactionAdder()
-
   // analytics
   useEffect(() => {
     ReactGA.pageview(window.location.pathname + window.location.search)
@@ -887,31 +720,6 @@ export default function ExchangePage({ initialCurrency }) {
           {t('highSlippageWarning')}
         </div>
       )}
-      <>
-        {account && (
-          <>
-            <>
-              <p className="orders-title">{`${t('Orders')} ${orders.length > 0 ? `(${orders.length})` : ''}`}</p>
-              {loading && (
-                <>
-                  <SpinnerWrapper src={Circle} alt="loader" /> Loading ...
-                  <br />
-                  <br />
-                </>
-              )}
-              {orders.length === 0 && !loading && <p>{t('noOpenOrders')}</p>}
-              {
-                <div>
-                  {orders.map(order => (
-                    <OrderCard key={order.witness} data={order} />
-                  ))}
-                </div>
-              }
-            </>
-            {/* <OrdersHistory /> */}
-          </>
-        )}
-      </>
     </>
   )
 }
