+++ src/pages/Swap/index.js
@@ -1,6 +1,10 @@
 import React from 'react'
 import ExchangePage from '../../components/ExchangePage'
+import Orders from '../../components/Orders'
 
 export default function Swap({ initialCurrency }) {
-  return <ExchangePage initialCurrency={initialCurrency} />
+  return <>
+    <ExchangePage initialCurrency={initialCurrency} />
+    <Orders />
+  </>
 }
